Move simulation solver from base simulation to PDE simulation (#1582)
#### Summary
Move `solver` attribute to the `BasePDESimulation`.


#### What does this implement/fix?
Not all simulations need `solver`, leading to awkward setting of
`solver=None` on `LinearSimulation`, and also on the layered 1D
simulations. This move the `solver` attribute to the `BasePDESimulation`
(the ones that actually need a solver).
