+++ simpeg/dask/simulation.py
@@ -53,8 +53,6 @@ def __init__(
     self,
     mesh=None,
     survey=None,
-    solver=None,
-    solver_opts=None,
     sensitivity_path="./sensitivity/",
     counter=None,
     verbose=False,
@@ -67,8 +65,6 @@ def __init__(
         self,
         mesh=mesh,
         survey=survey,
-        solver=solver,
-        solver_opts=solver_opts,
         sensitivity_path=sensitivity_path,
         counter=counter,
         verbose=verbose,
